IA-52 Delete Medicine Validate-Input Before Deletion (#95)
* IA-52 Delete Medicine Validate-Input Before Deletion

* IA-50 Delete Inventory (#94)

* Merge branch 'ia-52-delete-medicine-validate-input-before-deletion' of https://github.com/InfoTech501/infirmary-desktop-app into ia-52-delete-medicine-validate-input-before-deletion

* Merge branch 'ia-52-delete-medicine-validate-input-before-deletion' of https://github.com/InfoTech501/infirmary-desktop-app into ia-52-delete-medicine-validate-input-before-deletion

* IA-54 Add New Medicine – Input Validation for Name and Description (#96)

* IA-54 Add New Medicine – Input Validation for Name and Description

* IA-50 Delete Inventory (#94)

* Merge branch 'ia-54-add-new-medicine-input-validation-for-name-and-description' of https://github.com/InfoTech501/infirmary-desktop-app into ia-54-add-new-medicine-input-validation-for-name-and-description

---------


* IA-52 Delete Medicine Validate-Input Before Deletion

---------

+++ src/main/java/com/rocs/infirmary/desktop/InfirmarySystemApplication.java
@@ -382,27 +382,38 @@ public class InfirmarySystemApplication {
             case 9: {
                 try {
                     scanner.nextLine();
-
                     MedicineInventoryFacade medicineInventoryFacade = new MedicineInventoryFacadeImpl();
                     System.out.println("Delete Medicine By Item Name : ");
                     String itemName = scanner.nextLine().trim();
 
                     if (itemName.isEmpty()) {
                         System.out.println("No data Detected");
+                        break;
+                    }
+                   if (itemName.matches("\\d+")) {
+                       System.out.println("Invalid input: Medicine must be a string");
+                       break;
+                   }
+                        if (itemName.matches(".*[a-zA-Z].*")){
+                            System.out.println("Invalid input: Medicine must be a string");
+                            break;
+
                     } else if (!medicineInventoryFacade.IsAvailable(itemName)) {
                         System.out.println("This medicine " + itemName + " " + "does not exist");
+                        break;
+
                     }
-                    String confirmationMessage = "Are you sure you want to delete this Medicine Item? \n This action cannot be undone. ";
-                    int confirmation = InfirmarySystemApplication.getUserConfirmation(scanner, confirmationMessage);
+                        String confirmationMessage = "Are you sure you want to delete this Medicine Item? \n This action cannot be undone. ";
+                        int confirmation = InfirmarySystemApplication.getUserConfirmation(scanner, confirmationMessage);
 
-                    if (confirmation == 1) {
-                        boolean success = medicineInventoryFacade.deleteMedicineByItemName(itemName);
-                        System.out.println(success ? "Successfully Deleted" : "Failed to Delete");
+                        if (confirmation == 1) {
+                            boolean success = medicineInventoryFacade.deleteMedicineByItemName(itemName);
+                            System.out.println(success ? "Successfully Deleted" : "Failed to Delete");
 
-                    } else {
-                        System.out.println("Cancel the deletion. ");
+                        } else {
+                            System.out.println("Cancel the deletion. ");
+
                     }
-
                 } catch (RuntimeException e) {
                     throw new RuntimeException(e);
                 }
